+++ src/controllers/areaController.js
@@ -54,6 +54,7 @@ exports.getAllAreas = async (req, res) => {
 
     return res.status(200).json({
       status: 'success',
+      results: areasList.length,
       data: { areasList },
     });
   } catch (err) {
