+++ src/routes/userRouter.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const userController = require('../controllers/userController');
-const { isAuth, isAdmin } = require('../utils/authenticator');
+const { isAuth } = require('../utils/isAuth');
+const { isAdmin } = require('../utils/isAdmin');
 
 // Profile & Settings Routes 
 // These routes allow access to an authenticated user's private data.
